Attempt to fix GitHubStatusPush
The failures of GitHubStatusPush to work got worse (not better) after upgrading buildbot to latest. My theory here (based on https://github.com/buildbot/buildbot/issues/4105 and https://github.com/buildbot/buildbot/pull/4347) is that we're seeing 'undefined behavior'; in particular, the docs for SingleBranchScheduler say:

- The behavior of [SingleBranchScheduler] is undefined, if treeStableTimer is set, and changes from multiple branches, repositories or codebases are accepted by the filter."

Previously we were in fact accepting changes from multiple codebases; this code predates my work on this, but I presume the goal was to trigger rebuilds for a PR if changes to the relevant LLVM branch were pushed (in addition to Halide itself). That seems unnecessary to me (LLVM churn isn't high enough to justify it), but maybe I misunderstand the intent.

Anyway, I'm hoping this will address the lingering issues we have.