Switch buildbots to use mostly CMake
Since the CMake rewrite, we've been doing a fair amount of redundant testing on the CMake/Make builds on the buildbots; this proposes to move almost all of the testbranch testing on the buildbots to use CMake instead of Make.

This is more of a proposal than a real PR; it assumes that we want to start considering CMake to be our 'primary' build system everywhere, with Make supported on a best-effort basis, which is not somethere there is consensus on; this is an effort to start the dialog.

Note also that we still have to use Make for distrib builds (CMake has some open issues with this on Windows IIRC), so, again, this is more an open-for-discussion thing.