Stop building Debug distros on Windows
When building Windows distros, stop building the Debug variants; it takes a long time, is almost useless, and we don't think anyone uses it.

Also, as a drive-by fix, remove the 'config' option from `get_env()`; the only env var that we expect to vary between debug and release is LLVM_CONFIG, which is only set for makefile builds.

(Note that this change means the buildbots never attempt to do any debug builds of any sort anymore, but still iterate thru 'configs' lists that are just 'Release'; this is deliberate for now, as I'd rather wait to remove that plumbing after we're certain that Debug is gone for good.)