Don't force rebuilds of testbranches from LLVM changes
We currently trigger rebuilds for PRs (testbranch) if the Halide source changes *or* if the LLVM source changes. This PR eliminates the latter (for testbranches only), as changes in LLVM rarely break PRs this way, but changes in LLVM happen frequently enough that this sucks up a lot of buildbot time. (There's also apparently some UB involved that we can dodge; whether or not this will improve reliability is not clear.)