Update LLVM_MAIN -> 13, add LLVM_RELEASE_12
LLVM main branch is now 13, so update our versioning. Add LLVM_RELEASE_12 to the list of branches so that we build nightlies, but don't use it otherwise just yet.

(We really need to update the build rules to have HALIDE_MAIN test against both LLVM_12 and LLVM_MAIN, but that can wait for a bit while we get the buildbots unstuck)