Make script more dumb-text-editor friendly (#145)
LLVM and HALIDE are too-generic names for data structures that are used frequently here; searching for usages of these in a plain text editor (vs Python-aware IDE) is tedious because the terms are used so often in comments, etc. Appended _BRANCHES to both of them to make this use case easier.