Skip clang-tools-extra on most buildbots (#217)
We need some of these tools when building for ASAN, but (1) that's currently disabled because the work was put on hold in favor of higher-priority work, and (2) even when enabled, would only be done on linux-x86-64. So skip it on the slower systems (e.g. arm32, which currently take > 12 hours to build LLVM16 from scratch...)