Revise halide-main builders to be halide-nightly builders
For a long time, our buildbot system has treated changes to Halide `main` the same way as changes to PR testbranches: any commit should trigger a rebuild right away (well, within a ~5 minute threshold).

This PR proposes to change this to build Halide `main` distributions on a nightly basis instead, much as we do for LLVM, for a few reasons:
- We don't really need to rebuild the distributions right away; it's incredibly rare for them to detect a breakage that wasn't already known.
- It's rare that we need distrib builds at all except when producing releases (at least as of right now, since we don't really offer the interim distrib binaries to the public)
- It's common for the current 'main' builders to grab a bunch of buildbots during the middle of the workday, even if there are a bunch of pending testbranch PRs that really should get priority

So, instead, this proposes a nightly approach. This PR proposes to start the nightlies builds at 9PM PST (sorry, rest-of-world). Note that the LLVM nightlies currently start at 11pm, so this means that (1) the Halide nightlies will deliberately use the LLVM nightlies from the previous day (on the assumption that build breakages will be more likely to have been ironed out), and (2) it should be rare that this doesn't give enough time for the Halide nightlies to complete before the LLVM nightlies start (since the 'nightlies' builds don't run the full test suite).