Constrain the halide-nightly builds (#263)
* Constrain the halide-nightly builds

We currently default to building 'nightlies' from all the HALIDE_BRANCHES; in practice, this is overkill, as (eg) neither the halide-15 nor halide-16 branch are getting daily changes. This becomes problematic for our slower buildbots (esp. the armbots) as they now often don't finish the 'nightlies' until mid-morning the next day. This change makes a separate list of branches to build nightlies for, defaulting to just the main branch. (If/when we need new nightlies for other branches, eg for a dot-release, we can edit them back in; this isn't optimal, but until/unless we get better buildbot performance, it's probably the best option.)

* Update master.cfg